Do not add a SCMRevisionState of type MultiSCMRevisionState to the MultiSCMRevisionState variable associated with the overal job.
When a MultiSCM.checkout() is done, the list of SCMs is iterated
over.  For each SCM in the list, we obtain the SCMRevisionState
for that SCM and pass it down to the checkout() method for each SCM.
For a Subversion checkout, the Subversion plugin
downcasts the revision revision state to SVNRevisionState.

Without this fix, a revision state of type MultiSCMRevisionState
was being passed down into the Subversion plugin.  This was
causing a ClassCastException to be thrown.

[FIXED JENKINS-26303]